requirements.txt
fastapi
uvicorn
python-dotenv
requests
supabase
pydantic